requirements.txt
django==3.2
djangorestframework==3.12
psycopg2==2.9.6
drf-yasg==1.21.5
pillow==9.5.0
django-filter==23.2
djangorestframework-simplejwt==5.2.2